refactor: Moved events in namespace structure.
