--- src/Tutor.Core/DomainModel/KnowledgeComponents/AssessmentEventHelp/SoughtChallengeHelp.cs
@@ -1,6 +1,0 @@
-﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.AssessmentEventHelp
-{
-    public abstract class SoughtChallengeHelp : SoughtHelp
-    {
-    }
-}
