--- src/Tutor.Core/DomainModel/KnowledgeComponents/AssessmentEventHelp/SoughtHints.cs
@@ -1,6 +1,0 @@
-﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.AssessmentEventHelp
-{
-    public class SoughtHints : SoughtChallengeHelp
-    {
-    }
-}
