--- src/Tutor.Core/DomainModel/KnowledgeComponents/AssessmentEventHelp/SoughtSolution.cs
@@ -1,6 +1,0 @@
-﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.AssessmentEventHelp
-{
-    public class SoughtSolution : SoughtChallengeHelp
-    {
-    }
-}
