+++ src/Tutor.Core/DomainModel/KnowledgeComponents/Events/AssessmentEventEvents/HelpEvents/SoughtChallengeHelp.cs
@@ -1,0 +1,6 @@
+﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.Events.AssessmentEventEvents.HelpEvents
+{
+    public abstract class SoughtChallengeHelp : SoughtHelp
+    {
+    }
+}
