+++ src/Tutor.Core/DomainModel/KnowledgeComponents/Events/AssessmentEventEvents/HelpEvents/SoughtSolution.cs
@@ -1,0 +1,6 @@
+﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.Events.AssessmentEventEvents.HelpEvents
+{
+    public class SoughtSolution : SoughtChallengeHelp
+    {
+    }
+}
