+++ src/Tutor.Core/DomainModel/KnowledgeComponents/Events/KnowledgeComponentEvents/KnowledgeComponentCompleted.cs
@@ -1,0 +1,6 @@
+﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.Events.KnowledgeComponentEvents
+{
+    public class KnowledgeComponentCompleted : KnowledgeComponentEvent
+    {
+    }
+}
