+++ src/Tutor.Core/DomainModel/KnowledgeComponents/Events/KnowledgeComponentEvents/KnowledgeComponentPassed.cs
@@ -1,0 +1,6 @@
+﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.Events.KnowledgeComponentEvents
+{
+    public class KnowledgeComponentPassed : KnowledgeComponentEvent
+    {
+    }
+}
