+++ src/Tutor.Core/DomainModel/KnowledgeComponents/Events/KnowledgeComponentEvents/KnowledgeComponentSatisfied.cs
@@ -1,0 +1,6 @@
+﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.Events.KnowledgeComponentEvents
+{
+    public class KnowledgeComponentSatisfied : KnowledgeComponentEvent
+    {
+    }
+}
