+++ src/Tutor.Core/DomainModel/KnowledgeComponents/Events/KnowledgeComponentEvents/SessionLifecycleEvents/SessionAbandoned.cs
@@ -1,0 +1,6 @@
+﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.Events.KnowledgeComponentEvents.SessionLifecycleEvents
+{
+    public class SessionAbandoned : SessionLifecycleEvent
+    {
+    }
+}
