+++ src/Tutor.Core/DomainModel/KnowledgeComponents/Events/KnowledgeComponentEvents/SessionLifecycleEvents/SessionLaunched.cs
@@ -1,0 +1,6 @@
+﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.Events.KnowledgeComponentEvents.SessionLifecycleEvents
+{
+    public class SessionLaunched : SessionLifecycleEvent
+    {
+    }
+}
