+++ src/Tutor.Core/DomainModel/KnowledgeComponents/Events/KnowledgeComponentEvents/SessionLifecycleEvents/SessionLifecycleEvent.cs
@@ -1,0 +1,8 @@
+﻿using Tutor.Core.DomainModel.KnowledgeComponents.Events.KnowledgeComponentEvents;
+
+namespace Tutor.Core.DomainModel.KnowledgeComponents.Events.KnowledgeComponentEvents.SessionLifecycleEvents
+{
+    public class SessionLifecycleEvent : KnowledgeComponentEvent
+    {
+    }
+}
