+++ src/Tutor.Core/DomainModel/KnowledgeComponents/Events/KnowledgeComponentEvents/SessionLifecycleEvents/SessionTerminated.cs
@@ -1,0 +1,6 @@
+﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.Events.KnowledgeComponentEvents.SessionLifecycleEvents
+{
+    public class SessionTerminated : SessionLifecycleEvent
+    {
+    }
+}
