--- src/Tutor.Core/DomainModel/KnowledgeComponents/KnowledgeComponentCompleted.cs
@@ -1,6 +1,0 @@
-﻿namespace Tutor.Core.DomainModel.KnowledgeComponents
-{
-    public class KnowledgeComponentCompleted : KnowledgeComponentEvent
-    {
-    }
-}
