--- src/Tutor.Core/DomainModel/KnowledgeComponents/KnowledgeComponentPassed.cs
@@ -1,6 +1,0 @@
-﻿namespace Tutor.Core.DomainModel.KnowledgeComponents
-{
-    public class KnowledgeComponentPassed : KnowledgeComponentEvent
-    {
-    }
-}
