--- src/Tutor.Core/DomainModel/KnowledgeComponents/KnowledgeComponentSatisfied.cs
@@ -1,6 +1,0 @@
-﻿namespace Tutor.Core.DomainModel.KnowledgeComponents
-{
-    public class KnowledgeComponentSatisfied : KnowledgeComponentEvent
-    {
-    }
-}
