--- src/Tutor.Core/DomainModel/KnowledgeComponents/Sessions/SessionLaunched.cs
@@ -1,6 +1,0 @@
-﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.Sessions
-{
-    public class SessionLaunched : SessionLifecycleEvent
-    {
-    }
-}
