--- src/Tutor.Core/DomainModel/KnowledgeComponents/Sessions/SessionTerminated.cs
@@ -1,6 +1,0 @@
-﻿namespace Tutor.Core.DomainModel.KnowledgeComponents.Sessions
-{
-    public class SessionTerminated : SessionLifecycleEvent
-    {
-    }
-}
