+++ src/Tutor.Web/Controllers/Plugin/PluginController.cs
@@ -3,7 +3,6 @@ using Microsoft.AspNetCore.Mvc;
 using System;
 using Tutor.Core.DomainModel.AssessmentEvents;
 using Tutor.Core.DomainModel.AssessmentEvents.Challenges;
-using Tutor.Core.DomainModel.KnowledgeComponents.AssessmentEventHelp;
 using Tutor.Core.LearnerModel;
 using Tutor.Web.Controllers.Domain.DTOs.AssessmentEvents.Challenge;
 using Tutor.Web.Controllers.Learners.DTOs;
